+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## v1.1.0
+- Removes debounce on validate and submit
 - Fixes infinite loop with `initialValues` in `useForm()`
 - Fixes display of children in `<Field>` when the attribute `options` is undefined and using a custom component
 - Fixes React warnings of duplicate keys with `options` in `<Field>`